Port ccemux API and custom ROM to built-in plugin
