+++ build.gradle
@@ -64,23 +64,6 @@ dependencies {
 	testCompile 'junit:junit:4.12'
 }
 
-task customRom (type: Zip) {
-	from 'src/main/rom'
-
-	include '*'
-	include '*/*'
-
-	archiveName 'custom.rom'
-	destinationDir = file('src/main/resources')
-}
-
-configure(customRom) {
-	group = "Build"
-	description = "Packs custom rom files into an embeddable archive."
-}
-
-processResources.dependsOn customRom
-
 jar {
 	manifest {
 		attributes 'SplashScreen-Image': 'splash1.png', 'Implementation-Version': version
