+++ src/main/resources/META-INF/services/net.clgd.ccemux.plugins.Plugin
@@ -2,3 +2,4 @@ net.clgd.ccemux.plugins.builtin.AWTPlugin
 net.clgd.ccemux.plugins.builtin.CCTweaksPlugin
 net.clgd.ccemux.plugins.builtin.TRoRPlugin
 net.clgd.ccemux.plugins.builtin.HDFontPlugin
+net.clgd.ccemux.plugins.builtin.CCEmuXAPI
