+++ src/main/resources/ccemux.json
@@ -8,7 +8,6 @@
 	"termWidth": 51,
 	"termHeight": 19,
 	"termScale": 3,
-	"apiEnabled": true,
 	"renderer": "AWT",
 
 	"maxComputerCapacity": 1000000,
